Object Brick Definition - static data getter should return null instead of throwing an exception - see #3343

+++ models/DataObject/Objectbrick/Dao.php
@@ -40,9 +40,7 @@ class Dao extends Model\DataObject\Fieldcollection\Dao
         $values = [];
 
         foreach ($fieldDef->getAllowedTypes() as $type) {
-            try {
-                $definition = DataObject\Objectbrick\Definition::getByKey($type);
-            } catch (\Exception $e) {
+            if(!$definition = DataObject\Objectbrick\Definition::getByKey($type)) {
                 continue;
             }
 
@@ -133,14 +131,10 @@ class Dao extends Model\DataObject\Fieldcollection\Dao
         // this is to clean up also the inherited values
         $fieldDef = $object->getClass()->getFieldDefinition($this->model->getFieldname());
         foreach ($fieldDef->getAllowedTypes() as $type) {
-            try {
-                $definition = DataObject\Objectbrick\Definition::getByKey($type);
-            } catch (\Exception $e) {
-                continue;
+            if($definition = DataObject\Objectbrick\Definition::getByKey($type)) {
+                $tableName = $definition->getTableName($object->getClass(), true);
+                $this->db->delete($tableName, ['o_id' => $object->getId()]);
             }
-
-            $tableName = $definition->getTableName($object->getClass(), true);
-            $this->db->delete($tableName, ['o_id' => $object->getId()]);
         }
     }
 }
